requirements.txt
aiogram==3.22.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-dotenv==1.0.1
greenlet==3.1.1
redis==5.2.1

